Remove virtual functions which cause implicit mallocs making Fsm.h unsuitable for heap-free environments
Virtual functions cause classes to have a vtable.  This contains a pointer to type information for RTTI,
which is dynamically allocated in many C++ runtimes, and is needed for `dynamic_cast`.  This can be
avoided with the right compile flags but it is safer to avoid use of virtual functions altogether.